# simpler-api/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: philipp.schmurr@kit.edu
  description: "This is a Schema extraction API based on the OpenAPI 3.1 specification.\
    \  You can find out more about\nSwagger at [https://swagger.io](https://swagger.io). "
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Schema API - OpenAPI 3.1
  version: 0.7.0
...
servers:
- url: http://localhost:7373
...
paths:
...
  /schemata/{schemaId}/corrections/entities:
    post:
      description: Add an entirely new entity
      operationId: addNewEntityCorrection
      parameters:
      - $ref: '#/components/parameters/schemaId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Entity'
          application/x-yaml:
            schema:
              $ref: '#/components/schemas/Entity'
          application-xml:
            schema:
              $ref: '#/components/schemas/Entity'
        description: A new entity object
        required: true
      responses:
        "200":
          description: successful operation
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: bad request
        "404":
          description: schema not found
      summary: Add an entirely new entity
      tags:
      - correction
  /schemata/{schemaId}/corrections/entities/{entityId}:
    post:
      description: change entity name
      operationId: changeEntityNameCorrection
      parameters:
      - $ref: '#/components/parameters/schemaId'
      - $ref: '#/components/parameters/entityId'
      requestBody:
        content:
          text/plain:
            schema:
              type: string
        description: new name as plain string
        required: true
      responses:
        "200":
          description: successful operation
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: bad request
        "404":
          description: schema or entity not found
      summary: change entity name
      tags:
      - correction
  /schemata/{schemaId}/corrections/entities/{entityId}/modifiers:
    put:
      description: Update entity modifiers
      operationId: updateEntityModifiersCorrection
      parameters:
      - $ref: '#/components/parameters/schemaId'
      - $ref: '#/components/parameters/entityId'
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/EntityModifier'
              type: array
          application/x-yaml:
            schema:
              items:
                $ref: '#/components/schemas/EntityModifier'
              type: array
          application-xml:
            schema:
              items:
                $ref: '#/components/schemas/EntityModifier'
              type: array
              xml:
                name: EntityModifiers
        description: list of EntityModifier objects
        required: true
      responses:
        "200":
          description: successful operation
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: bad request
        "404":
          description: schema or entity not found
      summary: Update entity modifiers
      tags:
      - correction
  /schemata/{schemaId}/corrections/entities/{entityId}/attributes:
    post:
      description: Add an entirely new attribute
      operationId: addNewAttributeCorrection
      parameters:
      - $ref: '#/components/parameters/schemaId'
      - $ref: '#/components/parameters/entityId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Attribute'
          application/x-yaml:
            schema:
              $ref: '#/components/schemas/Attribute'
          application-xml:
            schema:
              $ref: '#/components/schemas/Attribute'
        description: A new attribute object
        required: true
      responses:
        "200":
          description: successful operation
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: bad request
        "404":
          description: schema or entity not found
      summary: Add an entirely new attribute
      tags:
      - correction
  /schemata/{schemaId}/corrections/entities/{entityId}/attributes/{attributeId}:
    post:
      description: change attribute name
      operationId: changeAttributeNameCorrection
      parameters:
      - $ref: '#/components/parameters/schemaId'
      - $ref: '#/components/parameters/entityId'
      - $ref: '#/components/parameters/attributeId'
      requestBody:
        content:
          text/plain:
            schema:
              type: string
        description: new name as plain string
        required: true
      responses:
        "200":
          description: successful operation
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: bad request
        "404":
          description: "schema, entity or attribute not found"
      summary: change attribute name
      tags:
      - correction
  /schemata/{schemaId}/corrections/entities/{entityId}/attributes/{attributeId}/modifiers:
    put:
      description: Update attribute modifiers
      operationId: updateAttributeModifiersCorrection
      parameters:
      - $ref: '#/components/parameters/schemaId'
      - $ref: '#/components/parameters/entityId'
      - $ref: '#/components/parameters/attributeId'
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/AttributeModifier'
              type: array
          application/x-yaml:
            schema:
              items:
                $ref: '#/components/schemas/AttributeModifier'
              type: array
          application-xml:
            schema:
              items:
                $ref: '#/components/schemas/AttributeModifier'
              type: array
              xml:
                name: AttributeModifiers
        description: list of AttributeModifier objects
        required: true
      responses:
        "200":
          description: successful operation
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: bad request
        "404":
          description: "schema, entity or attribute not found"
      summary: Update attribute modifiers
      tags:
      - correction
  /schemata/{schemaId}/corrections/entities/{entityId}/relations:
    post:
      description: Add an entirely new relation
      operationId: addNewRelationCorrection
      parameters:
      - $ref: '#/components/parameters/schemaId'
      - $ref: '#/components/parameters/entityId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Relation'
          application/x-yaml:
            schema:
              $ref: '#/components/schemas/Relation'
          application-xml:
            schema:
              $ref: '#/components/schemas/Relation'
        description: A new relation object
        required: true
      responses:
        "200":
          description: successful operation
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: bad request
        "404":
          description: schema or entity not found
      summary: Add an entirely new relation
      tags:
      - correction
  /schemata/{schemaId}/corrections/entities/{entityId}/relations/{relationId}:
    post:
      description: change relation name
      operationId: changeRelationNameCorrection
      parameters:
      - $ref: '#/components/parameters/schemaId'
      - $ref: '#/components/parameters/entityId'
      - $ref: '#/components/parameters/relationId'
      requestBody:
        content:
          text/plain:
            schema:
              type: string
        description: new name as plain string
        required: true
      responses:
        "200":
          description: successful operation
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: bad request
        "404":
          description: "schema, entity or relation not found"
      summary: change relation name
      tags:
      - correction
  /schemata/{schemaId}/corrections/entities/{entityId}/relations/{relationId}/modifiers:
    put:
      description: Update relation modifiers
      operationId: updateRelationModifiersCorrection
      parameters:
      - $ref: '#/components/parameters/schemaId'
      - $ref: '#/components/parameters/entityId'
      - $ref: '#/components/parameters/relationId'
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/RelationModifier'
              type: array
          application/x-yaml:
            schema:
              items:
                $ref: '#/components/schemas/RelationModifier'
              type: array
          application-xml:
            schema:
              items:
                $ref: '#/components/schemas/RelationModifier'
              type: array
              xml:
                name: RelationModifiers
        description: list of RelationModifier objects
        required: true
      responses:
        "200":
          description: successful operation
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: bad request
        "404":
          description: "schema, entity or relation not found"
      summary: Update relation modifiers
      tags:
      - correction
...
components:
  parameters:
...
    entityId:
      description: ID of entity to access
      explode: false
      in: path
      name: entityId
      required: true
      schema:
        type: string
      style: simple
    schemaId:
      description: ID of schema to access
      explode: false
      in: path
      name: schemaId
      required: true
      schema:
        type: string
      style: simple
    attributeId:
      description: ID of attribute to access
      explode: false
      in: path
      name: attributeId
      required: true
      schema:
        type: string
      style: simple
    relationId:
      description: ID of relation to access
      explode: false
      in: path
      name: relationId
      required: true
      schema:
        type: string
      style: simple
  schemas:
...
    AttributeModifier:
      enum:
      - AttributeModifier: key
      example:
        attributeModifier: attributeModifier
      properties:
        attributeModifier:
          type: string
      required:
      - attributeModifier
      title: AttributeModifier
    Attribute:
      example:
        attributeName:
        - attributeName
        - attributeName
        hasAttributeModifier:
        - attributeModifier: attributeModifier
        - attributeModifier: attributeModifier
      properties:
        hasAttributeModifier:
          default: []
          items:
            $ref: '#/components/schemas/AttributeModifier'
          title: hasAttributeModifier
          type: array
        attributeName:
          items:
            type: string
          minItems: 1
          title: attributeName
          type: array
      required:
      - attributeName
      title: Attribute
    EntityModifier:
      enum:
      - EntityModifier: weak
      example:
        entityModifier: entityModifier
      properties:
        entityModifier:
          type: string
      required:
      - entityModifier
      title: EntityModifier
    RelationModifier:
      enum:
      - RelationModifier: identifying
      example:
        relationModifier: relationModifier
      properties:
        relationModifier:
          type: string
      required:
      - relationModifier
      title: RelationModifier
    Cardinality:
      enum:
      - Cardinality: oneOrNone
      - Cardinality: oneOrMore
      - Cardinality: any
      - Cardinality: exactlyOne
      example:
        cardinality: cardinality
      properties:
        cardinality:
          type: string
      required:
      - cardinality
      title: Cardinality
    Relation:
      example:
        objectCardinality:
          cardinality: cardinality
        subjectCardinality:
          cardinality: cardinality
        hasSubjectEntity: ""
        relationName:
        - relationName
        - relationName
        hasObjectEntity: ""
        inverseRelation: inverseRelation
        hasAttribute:
        - attributeName:
          - attributeName
          - attributeName
          hasAttributeModifier:
          - attributeModifier: attributeModifier
          - attributeModifier: attributeModifier
        - attributeName:
          - attributeName
          - attributeName
          hasAttributeModifier:
          - attributeModifier: attributeModifier
          - attributeModifier: attributeModifier
        hasRelationModifier:
        - relationModifier: relationModifier
        - relationModifier: relationModifier
      properties:
        hasAttribute:
          default: []
          items:
            $ref: '#/components/schemas/Attribute'
          title: hasAttribute
          type: array
        hasObjectEntity:
          title: hasObjectEntity
        hasRelationModifier:
          default: []
          items:
            $ref: '#/components/schemas/RelationModifier'
          title: hasRelationModifier
          type: array
        hasSubjectEntity:
          title: hasSubjectEntity
        inverseRelation:
          nullable: true
          title: inverseRelation
          type: string
        objectCardinality:
          $ref: '#/components/schemas/Cardinality'
        subjectCardinality:
          $ref: '#/components/schemas/Cardinality'
        relationName:
          items:
            type: string
          minItems: 1
          title: relationName
          type: array
      required:
      - hasObjectEntity
      - hasSubjectEntity
      - objectCardinality
      - relationName
      - subjectCardinality
      title: Relation
    Entity:
      example:
        isObjectInRelation:
        - objectCardinality:
            cardinality: cardinality
          subjectCardinality:
            cardinality: cardinality
          hasSubjectEntity: ""
          relationName:
          - relationName
          - relationName
          hasObjectEntity: ""
          inverseRelation: inverseRelation
          hasAttribute:
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          hasRelationModifier:
          - relationModifier: relationModifier
          - relationModifier: relationModifier
        - objectCardinality:
            cardinality: cardinality
          subjectCardinality:
            cardinality: cardinality
          hasSubjectEntity: ""
          relationName:
          - relationName
          - relationName
          hasObjectEntity: ""
          inverseRelation: inverseRelation
          hasAttribute:
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          hasRelationModifier:
          - relationModifier: relationModifier
          - relationModifier: relationModifier
        hasEntityModifier:
        - entityModifier: entityModifier
        - entityModifier: entityModifier
        entityName:
        - entityName
        - entityName
        entityUrl: entityUrl
        hasAttribute:
        - attributeName:
          - attributeName
          - attributeName
          hasAttributeModifier:
          - attributeModifier: attributeModifier
          - attributeModifier: attributeModifier
        - attributeName:
          - attributeName
          - attributeName
          hasAttributeModifier:
          - attributeModifier: attributeModifier
          - attributeModifier: attributeModifier
        isSubjectInRelation:
        - objectCardinality:
            cardinality: cardinality
          subjectCardinality:
            cardinality: cardinality
          hasSubjectEntity: ""
          relationName:
          - relationName
          - relationName
          hasObjectEntity: ""
          inverseRelation: inverseRelation
          hasAttribute:
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          hasRelationModifier:
          - relationModifier: relationModifier
          - relationModifier: relationModifier
        - objectCardinality:
            cardinality: cardinality
          subjectCardinality:
            cardinality: cardinality
          hasSubjectEntity: ""
          relationName:
          - relationName
          - relationName
          hasObjectEntity: ""
          inverseRelation: inverseRelation
          hasAttribute:
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          hasRelationModifier:
          - relationModifier: relationModifier
          - relationModifier: relationModifier
      properties:
        hasAttribute:
          default: []
          items:
            $ref: '#/components/schemas/Attribute'
          title: hasAttribute
          type: array
        hasEntityModifier:
          default: []
          items:
            $ref: '#/components/schemas/EntityModifier'
          title: hasEntityModifier
          type: array
        isObjectInRelation:
          default: []
          items:
            $ref: '#/components/schemas/Relation'
          title: isObjectInRelation
          type: array
        isSubjectInRelation:
          default: []
          items:
            $ref: '#/components/schemas/Relation'
          title: isSubjectInRelation
          type: array
        entityName:
          items:
            type: string
          minItems: 1
          title: entityName
          type: array
        entityUrl:
          nullable: true
          title: entityUrl
          type: string
      required:
      - entityName
      title: Entity
...